update : post, user, comment 데이터 가져오기 완료
